Fix "caching" issue with USE_DEBUG_MALLOC CMake option
The original intent was that for debug builds, this option would be
turned on, and otherwise be turned off. However, this was done by adding
the option twice, instead of adding it once and then setting it based on
the build type.

Note that this is still somewhat wrong and will not work with
multi-configuration generators; but at least it won't "cache" the value
incorrectly.