Replace use of `add_compile_defs` abstraction
The two CMake functions `add_compile_defs` and `target_compile_defs`
were unnecessary. They just wrapped the CMake built-in commands
`add_compile_definitions` and `target_compile_definitions` with an extra
parameter, languages, which was then added in a generator expression for
every other argument. This is usually pointless, however, as CMake
natively understands the difference between a C library (with `.c`
files) and a C++ library (with `.cpp` files). CMake will not compile a C
library with a C++ compiler. The only time this makes sense is for
`INTERFACE` options for a library that will be linked with both C and
C++ code (e.g. `oelibc`).