ECalls/OCalls with single-level deep-copy marshaling.
With this new marshaling scheme, all inputs are serialized to
an input buffer and send over to the other side which
writes out all the outputs into an output buffer.
The output buffer is deserialized by the caller.

Previously, pointer and arrays were marshalled by pointer
when an ecall is made. That is, the pointer to the array/buffer
was put in the marshaling struct which was sent to the enclave.
The enclave read these pointers to host memory and then copied
over these array/buffer into buffers in enclave memory.
This kind of marshaling assumes that the enclave can access host
memory. Since the host cannot access enclave memory, when an
ocall was made, complete marshaling of all parameters to host
memory was done, similar in essence to the scheme below.

The new marshaling implemented by this changelist is more RPC style.
The contents (as opposed to pointers to) array and buffer inputs
are serialized to an input buffer. The input buffer contains
the marshaling struct first, followed by the serialized array/buffer
inputs. The enclave receives the buffer and clones it into enclave memory
in one swoop. It then patches up the pointers in the argument struct
(which is the first item in the cloned input buffer) to corresponding
pointers in the cloned memory buffer. Similarly all the ouputs are
serialized into an output buffer. For simplicity, the first item in the
output buffer is again the marshaling struct (though we just need the
ocall result and the return value fields of the struct).
The output parameters are then copied over from appropriate offsets
within the output buffer.
The same scheme is used for ecalls and ocalls.
For ocalls however the enclave allocates the input/output buffers directly
in host memory and the host directly uses them rather than copying.