Replace us of `if_supported` in libcxx/thrd tests
By removing the use of this module, we've eliminated multiple
invocations of the compiler for every test executable. The module
instructed CMake to test each flag for support, for every library, and
did not cache the values. This speeds up the tests from about 3 hours to
about 30 minutes.

Also replaced the hard-coded `std=c++14` with `CXX_STANDARD` properties,
and some various clean ups (but not all).